iot/Headcount/requirements.txt
ultralytics
opencv-python
flask
paho-mqtt
python-dotenv